Further investigation of Bully behaviour: * Repaired incorrect instance # printout (was delay instead of instance #)
* Add printouts in LeaderSelection.java -- printing out the TMan partners
before starting monitoring;
tried to look for partitions, but seems like there is a pretty good overlap
between the neighbors (at least when the PEER_TRUNCATE_VARIABLE=8)

* Changed the instance counter in LeaderSelection.java to increase after
each NewInstance event is triggered, instead of before each CurrentLeaderEvent
(since CurrentLeaderEvent is reported multiple times for the same leader and the
instance counter grew without any need)

* Changed the leader variable to static

* Final observation - when PEER_TRUNCATE_VARIABLES= 100 (i.e. we don't truncate)
the leader is elected ONCE (CurrentLeaderEvent is triggered by all neighbors)
and is stable in the converged sample.

Question: perhaps we need to make sure we propagate the leader across all nodes
once it's elected?